Moved the database classes to separate files
